Change destination of contact us button on landing page
