+++ core/templates/core/landing-page.html
@@ -49,7 +49,7 @@
             {{ page.proposition_text|safe }}
         </div>
         <div class="column-one-half button-container bidi-rtl">
-            <a href="{{ directory_components_urls.contact_us }}" class="button">{{ page.call_to_action_text }}</a>
+            <a href="{% url 'sector-list-cms-contact' slug=page.breadcrumbs.industrycontactpage.slug %}" class="button">{{ page.call_to_action_text }}</a>
         </div>
     </div>
 </section>
